move common code in AssignmentRepo and GroupAssignmentRepo to a mixin, also moved #disabled? method to view model
