+++ app/models/assignment_repo.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class AssignmentRepo < ApplicationRecord
+  include AssignmentRepoable
+
   update_index("assignment_repo#assignment_repo") { self }
 
   # TODO: remove this enum (dead code)
@@ -14,9 +16,6 @@ class AssignmentRepo < ApplicationRecord
 
   validates :assignment, presence: true
 
-  validates :github_repo_id, presence:   true
-  validates :github_repo_id, uniqueness: true
-
   validate :assignment_user_key_uniqueness
 
   # TODO: Remove this dependency from the model.
@@ -25,30 +24,7 @@ class AssignmentRepo < ApplicationRecord
   delegate :creator, :starter_code_repo_id, to: :assignment
   delegate :github_user,                    to: :user
   delegate :default_branch, :commits,       to: :github_repository
-
-  # This should really be in a view model
-  # but it'll live here for now.
-  def disabled?
-    @disabled ||= !github_repository.on_github? || !github_user.on_github?
-  end
-
-  def private?
-    !assignment.public_repo?
-  end
-
-  def github_team_id
-    repo_access.present? ? repo_access.github_team_id : nil
-  end
-
-  def github_repository
-    @github_repository ||= GitHubRepository.new(organization.github_client, github_repo_id)
-  end
-
-  def import_status
-    return "No starter code provided" unless assignment.starter_code?
-
-    github_repository.import_progress.status.humanize
-  end
+  delegate :github_team_id,                 to: :repo_access, allow_nil: true
 
   # Public: This method is used for legacy purposes
   # until we can get the transition finally completed
