+++ spec/models/assignment_repo_spec.rb
@@ -53,6 +53,12 @@ RSpec.describe AssignmentRepo, type: :model do
 
       expect(repo.valid?).to be_truthy
     end
+    it "expects github_team_id to be nil" do
+      repo = build(:assignment_repo, assignment: assignment, github_repo_id: 42)
+      repo.repo_access = nil
+
+      expect(repo.github_team_id).to be_nil
+    end
   end
 
   describe "#assignment_user_key_uniqueness" do
